+++ lib/config_default.rb
@@ -6,8 +6,6 @@ require "config_default/version"
 require "config_default/config"
 require "config_default/init"
 require "config_default/struct"
-require "config_default/rails_application_extension"
-require "config_default/rails_application_configuration_extension"
 
 module ConfigDefault
   extend self
@@ -20,8 +18,8 @@ module ConfigDefault
     yield(config) if block_given?
   end
 
-  def init_rails_monkey_patch!
-    ConfigDefault::Init.init_rails_monkey_patch!
+  def init_rails_patch!
+    ConfigDefault::Init.init_rails_patch!
   end
 
   def load(name, key: Rails.env, symbolize_keys: false, deep_symbolize_keys: false)
