--- lib/config_default/rails_application_extension.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-module ConfigDefault::RailsApplicationExtension
-  def config_for(name, env: Rails.env)
-    data = ConfigDefault.load(name, key: env, deep_symbolize_keys: true)
-    ActiveSupport::OrderedOptions.new.merge(data)
-  end
-end
