+++ lib/config_default/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module ConfigDefault
-  VERSION = "0.2.5"
+  VERSION = "0.3.0"
 end
